requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator>=2.0.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10

# File Processing
PyPDF2==3.0.1
pdfplumber==0.11.0
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==11.0.0
reportlab==4.0.9
fpdf2>=2.7.0
markdown-it-py>=3.0.0
matplotlib==3.8.2

# LLM & AI
openai==1.54.4
anthropic>=0.34.0
langchain==0.3.7
langchain-openai==0.2.8
tiktoken==0.8.0

# Vector Database (RAG)
pgvector>=0.2.0
numpy>=1.26.0,<2.0.0
sentence-transformers>=2.2.0

# AWS & Storage
boto3>=1.34.0
botocore>=1.34.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1

# OAuth
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
cryptography==42.0.0  # For token encryption

# Utilities
httpx==0.27.2
python-dateutil==2.9.0
paramiko>=3.0.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.0
